refactor(rust): split the extractors, and state `use` bindings as a query
rust_node_extractors.rs held two unrelated types and interleaved them:
RustDefinitionExtractor's impl blocks sat at lines 20, 50 and 128, then the whole
usage extractor, then more of the definition extractor at 790. Reading either one
meant skipping the other. Four modules now each answer one question:

  definition_extractor  what a file declares
  pattern_bindings      what a pattern binds, as opposed to reads
  usage_extractor       what a file reads
  self_type             `Self`, resolved to the type it stands for

extract_import_definition went with them, into five lines of definitions.scm. It
was 87 lines of nested traversal building the same ImportDefinition five times
over, and every case it decided was pure shape — the last segment of a path, an
item in a braced list, an alias. Stating it as a query also covers the nesting the
traversal did not reach, such as an alias inside a braced list.

Behaviour is unchanged either way, which the accuracy harness confirms: 743 of 743,
precision and recall 1.000, identical to the recorded baseline.
