+++ crates/core/src/languages/rust/mod.rs
@@ -1,7 +1,10 @@
 pub mod binding_queries;
+pub mod definition_extractor;
 pub mod definition_queries;
 pub mod dependency_resolver;
 pub mod format_string;
 pub mod formatter;
-pub mod rust_node_extractors;
+pub mod pattern_bindings;
 pub mod scope_queries;
+pub mod self_type;
+pub mod usage_extractor;
